+++ changelog.md
@@ -2,6 +2,6 @@
 ---
 
 - Changed storage to use JSON instead of SQLite 
-- Added a config with options to toggle layers and change options like max area size
+- Added a config with options to toggle layers and change options like max area size and z-index
   - The config can be reloaded alongside Pl3xMap using the `/pl3xmap reload` command
 - Added feedback in the form of action messages and sounds when placing/removing markers or points